agregando entidad de facturas

+++ src/main/java/com/gualteros/weaponsStore/repository/ItemFacturaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.gualteros.weaponsStore.models.ItemFactura;
 
-public interface ItemFacturaRepository extends JpaRepository<ItemFactura, Integer> {
+public interface ItemFacturaRepository extends JpaRepository<ItemFactura, Long> {
 
 }
